+++ packages/next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vexcms/admin-next
 
+## 0.1.0-alpha.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @vexcms/react@0.1.0-alpha.11
+  - @vexcms/core@0.1.0-alpha.11
+
 ## 0.1.0-alpha.10
 
 ### Patch Changes