+++ packages/richtext-plate/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vexcms/richtext
 
+## 0.1.0-alpha.11
+
+### Patch Changes
+
+- @vexcms/core@0.1.0-alpha.11
+
 ## 0.1.0-alpha.10
 
 ### Patch Changes